Refactor: controller단순화 및 service 생성으로 계층 분리
